+++ src/lib/api/upload/upload.ts
@@ -73,8 +73,14 @@ export async function handleFile({
       },
     });
 
+    const fileOverwrite = req.headers['x-zipline-file-overwrite'] === 'true';
+
+    if (fileOverwrite && !config.files.fileOverwrite) {
+      throw 'File overwrite is disabled in the configuration';
+    }
+
     if (existing) {
-      if (config.files.fileOverwrite) {
+      if (fileOverwrite && config.files.fileOverwrite) {
         await prisma.file.delete({
           where: {
             id: existing.id,
